+++ spec/dummy/app/models/my_model.rb
@@ -1,0 +1,2 @@
+class MyModel < ActiveRecord::Base
+end